% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_icd9_pc}
\alias{example_data_icd9_pc}
\title{ICD-9 Procedure Code Example Data}
\format{
\subsection{\code{example_data_icd9_pc}}{

A data frame with 116 rows and 1 column:
\describe{
\item{icd9_pc}{ICD-9 procedure codes}
}
}
}
\usage{
example_data_icd9_pc
}
\description{
An example dataset of ICD-9 procedure codes.
}
\keyword{datasets}
