% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_description}
\alias{get_description}
\title{Get Description}
\usage{
get_description(codes, icdVer, code_type)
}
\arguments{
\item{codes}{A vector of ICD diagnosis or procedure codes.}

\item{icdVer}{A number, either 9 or 10, indicating the ICD version.}

\item{code_type}{A string, either "dg" or "pc," indicating the codes are diagnosis or procedure, respectively.}
}
\value{
A dataframe with the ICD code and its description.
}
\description{
Function to get description of a code.
}
