% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{icd_read_fwf}
\alias{icd_read_fwf}
\title{Import GEM files, ICD9, ICD10 code descriptions}
\usage{
icd_read_fwf(path, width, colNames, fileEncoding = "")
}
\arguments{
\item{path}{The file path to the file to be loaded.}

\item{width}{A numeric vector specifying widths for each of the columns.}

\item{colNames}{A character vector specifying names of the columns.}

\item{fileEncoding}{A character string specifying the file encoding of the fixed width file.}
}
\value{
A dataframe of the ICD file.
}
\description{
Reads in fixed with ICD file, keeping all columns as character variables, trimming white space off ends of each column
}
