% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{match_code}
\alias{match_code}
\title{Exact or partial match}
\usage{
match_code(codes, code_vec, match_method = c("exact", "prefix"))
}
\arguments{
\item{codes}{Vector of character string of code to search for}

\item{code_vec}{Vector of character strings to compare against}

\item{match_method}{A string, specifying the method for matching the codes \itemize{
\item \code{"exact"} matches the provided code exactly.
\item \code{"prefix"} treats the provided code as a prefix for matching.
}}
}
\value{
Boolean for whether a match was found.
}
\description{
Wrapper for whether to search for an exact match or a partial match
of a code in a vector of codes.
}
\keyword{Internal}
