% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{match_prefixes}
\alias{match_prefixes}
\title{Match Prefixes}
\usage{
match_prefixes(strings, prefixes)
}
\arguments{
\item{strings}{A vector of character string to test}

\item{prefixes}{A vector of character strings representing the prefixes}
}
\value{
A boolean indicating whether the string matches any of the prefixes.
}
\description{
Checks whether a string matches any of a vector of prefixes.
}
\keyword{Internal}
