% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{forward_backward}
\alias{forward_backward}
\title{Forward-Backward Mapping}
\usage{
forward_backward(codes, icdVer_dest, code_type)
}
\arguments{
\item{codes}{A vector of ICD diagnosis or procedure codes.}

\item{icdVer_dest}{A number, either 9 or 10, indicating the destination ICD version.}

\item{code_type}{A string, either "dg" or "pc," indicating the codes are diagnosis or procedure, respectively.}
}
\value{
A dataframe with the bidirectionally mapped codes.
}
\description{
Performs the bidirectional mapping of a code.
}
\keyword{Internal}
