% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_code}
\alias{map_code}
\title{Map Code}
\usage{
map_code(
  codes,
  icdVer_dest,
  code_type = c("dg", "pc"),
  method = c("gem", "reverse-gem", "both", "multi-stage")
)
}
\arguments{
\item{codes}{A vector of ICD diagnosis or procedure codes.}

\item{icdVer_dest}{A number, either 9 or 10, indicating the destination ICD version.}

\item{code_type}{A string, either "dg" or "pc," indicating the codes are diagnosis or procedure, respectively.}

\item{method}{A string specifying the method for mapping the codes. The same methods as implemented by the \code{icd_convert} function in the touch \code{package}. \itemize{
\item \code{"gem"} performs a single forward mapping.
\item \code{"reverse-gem"} performs a single backward mapping.
\item \code{"both"} perfroms a single forward and backward mapping, combining the results.
\item \code{"multi-stage"} performs the multiple-stage mapping, as described in the \code{touch} package.
}}
}
\value{
A dataframe with the source code, the matching destination code, and additional columns for cases where multiple codes represent a single source code.
}
\description{
Map ICD diagnosis or procedure codes to and from either ICD9 or ICD10.
}
