% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_describe}
\alias{map_describe}
\title{Map Code and Get Description}
\usage{
map_describe(codes, icdVer_dest, code_type, keepMapCode = FALSE)
}
\arguments{
\item{codes}{A vector of ICD diagnosis or procedure codes.}

\item{icdVer_dest}{A number, either 9 or 10, indicating the destination ICD version.}

\item{code_type}{A string, either "dg" or "pc," indicating the codes are diagnosis or procedure, respectively.}

\item{keepMapCode}{Boolean - if true, returned data frame will keep the map code and associated columns, otherwise they are dropped.}
}
\value{
A dataframe with the original ICD code, the matching code, descriptions for both, and potentially columns for the map codes.
}
\description{
Function that maps code and gets description of the mapped codes
}
